add flask template filter
