+++ cubedash/summary/_stores.py
@@ -1674,25 +1674,6 @@ class SummaryStore:
             row.region_code,
         )
 
-    def get_dataset_center_time(self, dataset_id):
-        """
-        get dataset center_time from DATASET_SPATIAL table
-        to override the center_time value returned by datacube-core
-        """
-        rows = self._engine.execute(
-            select(
-                [
-                    DATASET_SPATIAL.c.center_time
-                ]
-            ).where(DATASET_SPATIAL.c.id == dataset_id)
-        ).fetchall()
-        if not rows:
-            return None
-        row = rows[0]
-
-        center_time = row.center_time
-        return center_time
-
 
 def _refresh_data(please_refresh: Set[PleaseRefresh], store: SummaryStore):
     """
